backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.3
pytest==8.3.2
gunicorn==23.0.0
flask-cors==4.0.1
